fix slide thumb selection
