+++ src/editor/SlideItemThumbEditor.tsx
@@ -3,12 +3,13 @@ import './SlideItemThumbEditor.scss';
 import { useState } from 'react';
 import { useSlideItemThumbSelecting } from '../event/SlideListEventListener';
 import { SlideItemThumbType } from './slideType';
-import { getValidSlideItemThumbSelected, useStateSettingNumber } from '../helper/helpers';
+import { useStateSettingNumber } from '../helper/helpers';
 import Tools from './Tools';
 import Editor from './Editor';
 import FlexResizer, { getPresentingFlexSize } from '../FlexResizer';
 import { parseHTML } from './slideParser';
 import { editorMapper } from './EditorBoxMapper';
+import { getValidSlideItemThumbSelected } from '../slide-presenting/SlideItemThumbList';
 
 export default function SlideItemThumbEditor() {
     const [slideItemThumb, setSlideItemThumb] = useState<SlideItemThumbType | null>(
